+++ client/src/components/home/AnimeContent.tsx
@@ -1,4 +1,5 @@
-import { EntyFragmentFragment, MediaStatus } from "@/generated/graphql";
+import { EntyFragmentFragment } from "@/generated/graphql";
+import { MediaStatus } from "@/lib/mediaStatus";
 import { AnimeSection } from "./AnimeSection";
 import { SearchBar } from "./SearchBar";
 import { ViewToggle } from "./ViewToggle";
