+++ client/src/pages/home.tsx
@@ -4,7 +4,7 @@ import { fetchUserAnime } from "@/lib/anilist";
 import { logger } from "@/lib/logger";
 import { LoadingView, ErrorAlert, AnimeContent } from "@/components/home";
 import { commonQueryOptions } from "@/lib/query-config";
-import { MediaListStatus } from "@/generated/graphql";
+import { MediaListStatus } from "@/lib/mediaListStatus";
 import { queryKeys } from "@/lib/queryKeys";
 
 export default function Home() {
